hcp365/hcp365-reporting/src/main/java/com/pulsepoint/hcp365/service/ReportTemplateColumnGroupService.java
package com.pulsepoint.hcp365.service;


import com.pulsepoint.hcp365.modal.ReportTemplateColumnGroup;
import com.pulsepoint.hcp365.repository.ReportTemplateColumnGroupRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ReportTemplateColumnGroupService {

    @Autowired
    ReportTemplateColumnGroupRepository repository;

    public List<ReportTemplateColumnGroup> getAll(){
        return repository.findByStatusOrderByOrdinal(true);
    }
}

